dwarf_parser: Allow more types in const values
Recent versions of `rustc` will emit `DW_FORM_ref_addr` and
`DW_FORM_block1` for const values and, whereas previously it would only
use `DW_FORM_ref_udata`.

Handle these types, and promote constants to a u128 to handle the larger
value.

+++ src/model.rs
@@ -621,7 +621,7 @@ pub struct SubParameter {
     /// constants.
     ///
     /// TODO: type probably needs to be more general.
-    pub const_value: Option<u64>,
+    pub const_value: Option<u128>,
     /// Location in debug info.
     pub offset: gimli::UnitSectionOffset,
 }
